fix: add decimal to exchange, fix decimal for create, update rate

+++ src/currency_exchange/dtos.py
@@ -36,13 +36,13 @@ class RatePostUpdateDto:
 class ExchangeDto:
     base_currency: CurrencyDto
     target_currency: CurrencyDto
-    rate: float
-    amount: float
-    converted_amount: float
+    rate: Decimal
+    amount: Decimal
+    converted_amount: Decimal
 
 
 @dataclass
 class ExchangePostDto:
     from_currency_code: str
     to_currency_code: str
-    amount: float
+    amount: Decimal
